Merged PR 690: Merge IES-181-DB-Work-Part-1 to PtmContracts
**Overview:**
Update PTM ProposalChecklist table. One column renamed, two new columns created. Save functionality now works for the latter two fields. The latter two fields already exist in the UI, the backend needed to be wired up. Profit/Fee and Com are now split on the frontend and backend. The former column will remain but will be unused for legacy purposes.

**Note:**
For this MR and branch we had the DBA create a new database to work with so that UAT would not be affected. The new database is called 'genTRAC_Contacts'. Notice it says "Contact" not "Contracts". The genTRAC model and unit test App.config were modified to connect to this database. As well as the web Web.config. When the PtmContracts branch is ready to be merged to UAT, the three aforementioned config files will need to point back to the UAT genTRAC DB. StoredProcs will need to be ran on the UAT database and the model may also need to be updated if auto-generated files don't play nicely.

**Specifics:**
- Modify database column 'ProfitFee' renamed to 'ProfitFeeWithCom' (unused, staying for legacy)
- Add database column 'Profit'
- Add database column 'Com'
- GenTRAC.Models updated accordingly
- GenTRAC .edmx updated accordingly
- Stored Procedures updated accordingly
- Proposal Checklist unit tests updated accordingly

**JIRA Task:**
_https://ebstools-jira.us.lmco.com/browse/IES-442_

**Screencap of the Database:**
![image.png](https://idetfs01.us.lmco.com/Estimating/ec33f0ce-06e6-4674-9d97-643c8eb6e517/_apis/git/repositories/c962790d-14a5-4524-9fe8-35b4cce33cc3/pullRequests/690/attachments/image.png)

**Screencap of the UI:**
![image.png](https://idetfs01.us.lmco.com/Estimating/ec33f0ce-06e6-4674-9d97-643c8eb6e517/_apis/git/repositories/c962790d-14a5-4524-9fe8-35b4cce33cc3/pullRequests/690/attachments/image%20(4).png)