Merged PR 742: Merge IES-177-Integration to PtmContracts
First and second part of the the PTM integration task. Added Contracts Controller Logic for PTM Contracts. With conversions for Model-to-DTO and DTO-to-Model. Added Unit Tests for ContractsControllerLogic as well.

NOTE: The SaveContract and Validation methods will need unit tests as well. The validate method still needs to be created. This will be done on Integration Part 3 or Save Work.