Merged PR 957: IES-1384: Unable to Save Text Entered into 2nd RTE Field (Basis Rationale)
User (May Ann Li) reported that she was unable to save the text she had entered into the IWTA IBOE, Basis of and Rationale for LM Space Proposed Value for IWTA field. I confirmed that I too was unable to save the text that I entered into the field.

Fixed issue by adding the last missing piece of the puzzle which was given the dto field the proper value :)