Merged PR 974: IES-1507: Change Cost - Cost through COM Middle Tier
Completed the Middle Tier portion of implementing the [IES-1409](https://ebstools-jira.us.lmco.com/browse/IES-1409) story.

Finished up the additional loader work, updated the data to get the new Cost through COM data and updated the unit tests.
Updated the DTO and View Model with new properties. Touched where it was being loaded, saved and converted.

Added a date to the web config for when the feature is supposed to start - will need to confirm with @<ACCT04\paliderd> when the ACTUAL date is, as of right now it is set to 7/21/2022 which is the Thursday this item is set into code review.

Added validation.