Merged PR 997: IES-1582: Wrap Up
While working on IES-1582 there were a couple cases that was breaking the original logic. One of them being the most common which was null exceptions. In this wrap up task I added logic for null checks as well as updated the unit tests to include more scenarios to increase robustness.

Here are screenshots of the test passing and it working from the IES Communications API pinging:

![CoverSheetDataLoaderTestsWorking.png](https://idetfs01.us.lmco.com/Estimating/ec33f0ce-06e6-4674-9d97-643c8eb6e517/_apis/git/repositories/c962790d-14a5-4524-9fe8-35b4cce33cc3/pullRequests/997/attachments/CoverSheetDataLoaderTestsWorking.png)

![CoverSheetDataLoaderQueryWorking.png](https://idetfs01.us.lmco.com/Estimating/ec33f0ce-06e6-4674-9d97-643c8eb6e517/_apis/git/repositories/c962790d-14a5-4524-9fe8-35b4cce33cc3/pullRequests/997/attachments/CoverSheetDataLoaderQueryWorking.png)

![image.png](https://idetfs01.us.lmco.com/Estimating/ec33f0ce-06e6-4674-9d97-643c8eb6e517/_apis/git/repositories/c962790d-14a5-4524-9fe8-35b4cce33cc3/pullRequests/997/attachments/image.png)