Merged PR 1219: PROPH-250: Updated email Text for RDM Publishing
**Priority - nice to have by 4/3, need to have within a week.**

Please add the following text to the email that is delivered to RDM users when a PPR&D is published.

"For Contract POCs, the PPR&D Disclosure Log can be used to create the necessary disclosures for CCoPD covered proposals.  The log will be updated to include the specifics of this revision shortly."

The above text should be added below the current email text.

Once added the full text should appear as shown below....

The PPR&D has been revised and is available in RDM.

Program Contracts and Estimating are expected to review the revised PPR&D in full and provide the appropriate disclosures for your CCOPD proposal(s) to your Contracting Officers. Disclosable information may include data beyond rates and factors such as CAS, system status and Proposal Disclosures. Please note that disclosure artifacts are to be kept in the Contracts file and tracked in the proposal disclosure log for proper records retention and audit purposes.

For Contract POCs, the PPR&D Disclosure Log can be used to create the necessary disclosures for CCoPD covered proposals.  The log will be updated to include the specifics of this revision shortly.

Screenshot:
![emailrdm.png](https://idetfs01/Estimating/ec33f0ce-06e6-4674-9d97-643c8eb6e517/_apis/git/repositories/c962790d-14a5-4524-9fe8-35b4cce33cc3/pullRequests/1219/attachments/emailrdm.png)