Merged PR 1517: PROPH-1608: Create Feature Flag API Controller
With the new Angular applications, we will need a way to hide items behind a configuration (feature flag). Previously, in other applications we utilized a `web.config` file or the `appsettings.json` file.

We will continue to use the `appsettings.json` file since the back-end is .NET 6 however we will need to pass down the configurations from the backend.

![Capture.PNG](https://idetfs01/Estimating/ec33f0ce-06e6-4674-9d97-643c8eb6e517/_apis/git/repositories/c962790d-14a5-4524-9fe8-35b4cce33cc3/pullRequests/1517/attachments/Capture.PNG)